+++ silhouette/app/com/mohiva/play/silhouette/impl/providers/oauth2/ClefProvider.scala
@@ -58,7 +58,7 @@ trait BaseClefProvider extends OAuth2Provider {
       (json \ "error").asOpt[String] match {
         case Some(errorMsg) =>
           throw new ProfileRetrievalException(SpecifiedProfileError.format(id, errorMsg))
-        case _ => profileParser.parse(json)
+        case _ => profileParser.parse(json, authInfo)
       }
     }
   }
@@ -67,15 +67,16 @@ trait BaseClefProvider extends OAuth2Provider {
 /**
  * The profile parser for the common social profile.
  */
-class ClefProfileParser extends SocialProfileParser[JsValue, CommonSocialProfile] {
+class ClefProfileParser extends SocialProfileParser[JsValue, CommonSocialProfile, OAuth2Info] {
 
   /**
    * Parses the social profile.
    *
    * @param json The content returned from the provider.
+   * @param authInfo The auth info to query the provider again for additional data.
    * @return The social profile from given result.
    */
-  override def parse(json: JsValue) = Future.successful {
+  override def parse(json: JsValue, authInfo: OAuth2Info) = Future.successful {
     val userID = (json \ "info" \ "id").as[Long].toString
     val firstName = (json \ "info" \ "first_name").asOpt[String]
     val lastName = (json \ "info" \ "last_name").asOpt[String]